requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
openai==1.13.3
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
python-telegram-bot==20.8
reportlab==4.1.0
requests==2.31.0
sniffio==1.3.1
SpeechRecognition==3.10.1
tqdm==4.66.2
typing==3.7.4.3
typing_extensions==4.10.0
urllib3==2.2.1
